Initial commit
Created a project with a library and an app module.

Created some of the core classes and packages.

Implemented context wrapping and view inflation interception.
All string related method calls reach TxResources, which at the
moment prefixes all strings with "test: ".

TxResources overrides some of the methods that return strings. The rest of
the string methods call internally the ones we have already overridden.
The remaining methods are delegated to the wrapped Resources object.

We avoid altering the value of Android package strings.
Returning an altered value, such as in the case of
"config_inputEventCompatProcessorOverrideClassName",
can result to crashes.

ViewTransformer classes take care of changing string related
setters on Android's views during the inflation process. Since
all views can potentially have a tooltip or content description,
we check if these attributes have been defined in all view types.

The Activity boilerplate code to establish the wrapping has been
added to the app's BaseActivity and BaseAppCompatActivity classes.
Special handling was needed for the use of:
"AppCompatDelegate.setCompatVectorFromResourcesEnabled(true)",
where "VectorEnabledTintResourcesWrapper" is used.