Parse CDS response with GSON
LocaleDate has been introduced so that the structure
of the classes that hold the locales are clearly defined in the
codebase.

Server's response is parsed into a `TxResponseData` object using
GSON.

To avoid passing around HashMaps, `TranslationMap` and `LocaleStrings`
classes have been introduced. They can be printed in a
human readable form and can be serialized or deserialzed
using GSON.