Parse and print CDS response when pushing source strings
CDSHandler has been updated so that the CDS's response is parsed
to the corresponding LocaleData object. Affected classes and their unit
tests have been updated.

Improvements to test-only resources in txsdk

The unit tests of txsdk require some resources. At the same time,
we don't want these resources to exist when building the app or
the android library. As a workaround we set these resources
in the debug build variant. The downside is that the unit tests
that run in release mode fail, because these resources are not
accessible in this case.

The updated txsdk gradle configuration adds the required
resources only when running a test. The resources have been
moved from the debug to the test directory. Some unit tests in
the clitool module accessed these resourced by path and
have now been updated to point to the new location.

Improvements in transitive dependencies

"clitool" module depends on "common" module, which depends
on other libraries. In order to build a fat .jar that contains all needed
libraries, we had to re-declare "common's libraries in "clitool".
Now, this is not needed as we use the "runtimeClasspath" when
building the fat .jar.