Added pull command in cli
CDSHandler has a low-level fetch method that
allows the caller to acces the translation input stream
from the server. The existing fetch method uses it under the hood.

TranslationsDownloader makes use of the new fetch method to
save the translations as files in the specified directory.

Added pull command in cli.

Moved token and secret option inside the subcommands.

Updated the cli section of the readme.

Added unit tests.