Save and and load TranslationMap to/from disk
Added TranslationMap in common submodule so that
we can save/load a TranslationMap to/from disk.

Added TranslationMapAndroid in txnative submodule that
extends TranslationMap so that it can read files from
an application's raw asset files using Android's AssetManager.

To test TranslationMap and TranslationMapAndroid on
an Android filesystem, instrumented tests have been
added in txnative submodule. Directories containing
translations have been added in the "assets" directory
of the "androidTest" sourceset and are used in the
instrumented unit tests.

Changed NativeCore so that it treats empty strings
as non existing translations.

Renamed Cache to TxCache and moved it to its own
package. Added get() method to return all translations.
Added unit tests.

Implemented equals() and hashCode() method for some
classes of LocaleData so that we can compare them.
Added unit tests for them.

Updated "txsdk"'s build gradle:
"common" module is now added as an api to "txsdk"
module, since we are probably going to expose some
classes to the end SDK user.