Full cache implementation
Moved exising cache files in their own  cache package.

Removed TxCache#getSuppotedLocales().

Added new internfaces and classes.

Created unit tests and intrumented tests. Added files at
androidTest/assets directory so that they are accessible
during instrumented testing.

Updated TxNative and NativeCore so that TxStandardCache
is used, instead of TxMemoryCache. Updated the documentation.
When NativeCore received a null or empty map from CDSHandler
it does not update the cache.

Added copy constructor for LocaleStrings and TranslationMap.

Changed TranslationMapStorage#fromDisk() behavior so that it
returns null instead of an empty map when no translations are found.
This is because distinguishing between no folder found and no
translations found is not that easy in all cases.

Updated the name of existing unit tests.

Updated readme.