Publish in Maven Central
The version number of common, txsdk and cli are defined in
the projet's build gradle as extra properties.

The cli gets its version number dynamically.

Created publish-helper.gradle file which is included
in the build.gradle of txsdk and common modules.
It is responsible for signing and publishing the modules to
Mavel repository along with their sources and javadoc.
The signing configuration and sonatype credentials can
be read from "local.properties" file or via environmental
variables.

Updated readme with gradle dependancy instructions.

Updated documentation to fix javadoc generation errors.

Added rules in txsdk consumer proguard file to solve
issues when the sdk is used in an app with R8/proguard
enabled.

A task was created  to generate a javadoc that combines
the sources of txsk and common submodules. The aggregatd
javadoc uses the doc/readme.html in its overview page.