Add release Github workflow
The workflow is triggered by new releases and performes these steps:
* it builds and uploads the CLI tool as an artifact to the release that triggered it
* it generates and deploys the documentation to Github pages and it builds and
 publishes the sdk to Maven.
If the tag ends with "SNAPSHOT", only publishing to Maven takes place.

Updated gradle:

io.github.gradle-nexus.publish-plugin is used to publish to maven
and also close and release the staging repository. It also handles
snapshot releases. The original maven repository configuration
in "publish-helper.gradle" has been removed  to avoid creating
2 similar publishing tasks. We want the plugin to handle that.

The extra properties required for signing and publishing has been
moved to the root "build.gradle". If 'PGP_KEY_CONTEXT" environmental
variable is found, it is written to a file and "signing.secretKeyRingFile" is
updated to point to the file location. This way the script can work with
Github CI, but it can also work when the key is stored locally.

Removed the CLI jar from the repo

Updated readme