Make Appcompat dependency optional
txsdk's build.gradle declared an "implementation" dependency to
'androidx.appcompat:appcompat:1.2.0'. The dependency is now changed to
"compile-only". This way, we are not adding Appcompat to an app that is not
already using Appcompat.

Special care was taken for accessing  Appcompat methods and classes, as
doing so in an app that is not using Appcompat will result in a crash, since the
Appcompat classes will not be found at runtime. Utils#isAppcompatPresent()
is used before accessing an Appcompat class.

Removing Appcompat also affects multilingual support since Appcompat adds some
strings which are localized in lots of locales. Readme was updated with extended
instructions in case Appcompat is not used.

'androidx.annotation:annotation' dependency has been changed from "implementation" to
"compileOnly", as it is only needed at compile time (and when building the javadoc).
Since "compileOnly" dependencies are not inherited by the test sourcesets, it had to be
redeclared using "testCompileOnly".