Change step order in release workflow
The step that publishes the SDK to Maven can fail for network
or other reasons. At the same time, the javadoc deployment to
Github pages is deterministic.

This is why, publishing to Maven was placed as a last step
in the  workflow. This way, even if it fails, we are certain
that the other steps succeeded. We can always
publish to Maven manually.

Updated readme.html to reflect the latest readme.md