Add x-native-sdk header in CDSHandler
The build.gradle of "common" module can now generate  a
BuildProperites.java file that exposes some gradle properties
to Java.

BuildProperties#getSDKVersion() is used in CDSHandler#addHeaders() so that
"x-native-sdk" -> "mobile/android/{TXNative.version}" header is added when pushing
or pulling strings.

CLI used to get the CLI version from the "Implementation-Version" property of the jar
file. We now use BuildProperties#getCLIVersion() to get the version.