Both the SDK and the CLI pull command can now pull source strings.
Updated readme.md and readme.html regarding these changes.

Updated the CLI command description regarding the source strings.

Updated CDSHanlder and CDSHanlderAndroid documentation regarding the source strings.

NativeCore configures CDSHandlerAndroid so that both the translated and the source locales
are downloaded. The translate methods can now serve source strings from the cache and
fall back to the Android provided source strings. The following helper methods have been
created for this purpose: getSourceString(), getSourceQuantityString(). They are used when
the current locale is the source locale and when feeding a source string to the MissingPolicy.

getLocalizedQuantityString() used to follow the plural rules of the current locale. Now it follows
the plural rules of the locale of the provided Resources instance. This way, the method can be
used for the source locale rules or the current locale rules. The method has been made static.

mDefaultResources has been removed as it is not needed.

TxResourcesh has a getWrappedResources() method

A warning has been added to Utils#getDefaultLocaleResources()

Several unit tests have been added to NativeCoreTest to test the usage of source strings
provided by the cache when the source locale is used and when source missing policy is used.
Previous tests have been updated due to the changes in NativeCore methods.