Update CDSHandler logic to v2
Updates CDSHandler logic so that it can work with the version 2
of the CDS documentation. Updates the push logic with the new job
oriented approach. Adds time limit exception when pushing source strings.

Updates the response data structures that the new CDS version produces (LocadeData).

Refactor CLI's MainClass to account for the updates in CDSHandler and LcoaleData.

Updates unit tests for the class and method refactoring.

Adds unit tests for the new job oriented approach on push.