Update to Gradle 7.2 and JDK 12
Updated Gradle to 7.2 and Android Gradle plugin to 7.0.2

This Gradle version requires JDK 11. JDK 11 has a bug when
generating the Javadoc, which results in broken search functionality.
This is why we will use JDK 12.

Updated Github workflows to use JDK 12.

Replaced jcenter with mavenCentral.

Added some arguments for javadoc generation so that it
executes correctly with the latest JDK.

Corrected two errors in the documentation.