Update to AppCompat 1.5.0
AppCompat has been updated to 1.5.0 in the SDK and the app Gradle files.

BaseActivity.java and BaseAppCompatActivity,java have been moved from the app
module to the SDK (txsdk) module. It's better to have these classes bundled in the
library instead of instructing our users to copy paste the code from the sample app.

BaseAppCompatActivity used to wrap the resources with `VectorEnabledTintResourcesWrapper`
to enable TxNative functionality when "setCompatVectorFromResourcesEnabled()" was used
and the app was running on older  platforms (< API 21).
AppCompat 1.5.0 does not need this workaround. This is why the respective code has
been commented out. A comment is left for users that use an older AppCompat version
and need the workaround.

Readme.md has been updated regarding the previous changes.

Some test methods have been added in MyApplication as commented out code.