Access test resources directly
In txsdk module, we have test resources that are used in local unit tests.

In order to make these resources visible to the unit tests, we used to add them
in the main source set. The downside of this approach is that the resources were
released in the final .aar.

We have removed this configuration from build.gradle and we now access the
resources via the R class that is generated for the test component of txsdk:
com.transifex.txnative.test.R