Pass wrapped resources in AndroidMissingPolicy
Previously, AndroidMissingPolicy was initialized with a context, whose resources
were used for getting a translated string.

Using the same context for all calls can have negative side-effects. For example,
if  the resources configurations is changed inside an activity for forcing a
locale, the context used in AndroidMissingPolicy will not be affected.

This is why we no longer provide a context for constructing AndroidMissingPolicy.
The original constructor has been deprecated but not removed. The provided
context is ignored.

The MissingPolicy interface was updated so that all methods provide a Resources
object.

The AndroidMissingPolicy uses the provided Resources object to retrieve
translated strings.

NativeCore passes the wrapped resources object to the missing policy. We don't
want to pass TxResources because calling an overriden method would result in a
recursive loop.

All classes and unit tests have been updated.

The material library has been updated to 1.7.0.

Readme was updated.