Update dependencies
Updated appcompat to 1.6.1

Moved the version number of some dependencies in the "versions" variable of the
main gradle file.

Did some small updates as indicated by IntelliJ:

* register tasks instead of creating new ones
* use `configureEach`
* remove `buildToolsVersion`
* migrate `compilesdkversion` to `compieSdk` inside the android bracket

Due to updating the gson library, a conflict occurred with
`META-INF/versions/9/module-info.class` which was already provided by the
jdom library. An "exclude" option has been added in the jar configuration to
solve this issue.