Re-enable Kotlin
Kotlin plugin and kotlin-stdlib were removed from the txsdk module in
11c91d22b1ec0f477687ee3573b9f2da66d5b5fb

The removal of the plugin resulted in kotlin files not being compiled and
bundled in the final library. The only kotlin file that the project has and
was affected, was: `TxContextWrappingDelegate`

Now, this class will be available to library users to use it if they want.