Change ViewPump dependency
Changed ViewPump from  'io.github.inflationx:viewpump:2.0.3' to  "dev.b3nedikt.viewpump:viewpump:4.0.14".

This version has a workaround for WebViews, as it provides them with the base
resources. This solves the original WebView crash and our workaround is no
longer needed. Removing our workaround solves the issue where a crash occurred
when a view with attributes was loaded after a WebView.

The new library has very similar API to the previous one, so very few changes
were required. It doesn't require wrapping the Context as the previous one.
So, `TxNative#wrap()` and `TxNative#generalWarp()` now have the same implementation.
The latter has now been deprecated. The new library requires wrapping the
`AppCompatDeledate`, in a similar fashion that we do with our
`TxContextWrappingDelegateJava`. This leaded to changes to
`TxBaseAppCompatActivity`. Previously we created a new instance of
`TxContextWrappingDelegateJava`, but this is no longer enough. In case a TxNative
user, uses `TxContextWrappingDelegateJava` directly, `TxNative` will not work
as expected after this update. To avoid a wrong configuration `TxContextWrappingDelegateJava`
has been renamed to `TxContextWrappingDelegateJava2`, which will lead to a
compilation error in such a case.

The new ViewPump library requires AppCompat. TxNative did not require it up
until now. So, some changes were necessary to move from supporting AppCompat
to requiring it:

* the appcompat dependency was changed from "compileOnly" to "implementation"
* documentation changes